backend/requirements.txt
mysql-connector-python
attrs
dataclasses
flask
flask-cors